Refactor db methods getPrevEntityVersion and getFrothyRegion.
